Extract shared code into RegularTable
