+++ h2/src/main/org/h2/mvstore/db/MVTable.java
@@ -7,11 +7,6 @@ package org.h2.mvstore.db;
 
 import java.util.ArrayDeque;
 import java.util.ArrayList;
-import java.util.Collections;
-import java.util.Comparator;
-import java.util.HashSet;
-import java.util.Set;
-import java.util.concurrent.ConcurrentHashMap;
 import java.util.concurrent.TimeUnit;
 import java.util.concurrent.atomic.AtomicInteger;
 import java.util.concurrent.atomic.AtomicLong;
@@ -19,8 +14,6 @@ import java.util.concurrent.atomic.AtomicLong;
 import org.h2.api.DatabaseEventListener;
 import org.h2.api.ErrorCode;
 import org.h2.command.ddl.CreateTableData;
-import org.h2.constraint.Constraint;
-import org.h2.constraint.ConstraintReferential;
 import org.h2.engine.Constants;
 import org.h2.engine.DbObject;
 import org.h2.engine.Session;
@@ -39,19 +32,15 @@ import org.h2.result.SearchRow;
 import org.h2.schema.SchemaObject;
 import org.h2.table.Column;
 import org.h2.table.IndexColumn;
-import org.h2.table.Table;
-import org.h2.table.TableBase;
-import org.h2.table.TableType;
+import org.h2.table.RegularTable;
 import org.h2.util.DebuggingThreadLocal;
 import org.h2.util.MathUtils;
 import org.h2.util.Utils;
-import org.h2.value.DataType;
-import org.h2.value.Value;
 
 /**
  * A table stored in a MVStore.
  */
-public class MVTable extends TableBase {
+public class MVTable extends RegularTable {
     /**
      * The table name this thread is waiting to lock.
      */
@@ -107,11 +96,6 @@ public class MVTable extends TableBase {
     private MVPrimaryIndex primaryIndex;
     private final ArrayList<Index> indexes = Utils.newSmallArrayList();
     private final AtomicLong lastModificationId = new AtomicLong();
-    private volatile Session lockExclusiveSession;
-
-    // using a ConcurrentHashMap as a set
-    private final ConcurrentHashMap<Session, Session> lockSharedSessions =
-            new ConcurrentHashMap<>();
 
     /**
      * The queue of sessions waiting to lock the table. It is a FIFO queue to
@@ -121,8 +105,6 @@ public class MVTable extends TableBase {
     private final Trace traceLock;
     private final AtomicInteger changesUntilAnalyze;
     private int nextAnalyze;
-    private final boolean containsLargeObject;
-    private Column rowIdColumn;
 
     private final MVTableEngine.Store store;
     private final TransactionStore transactionStore;
@@ -133,15 +115,6 @@ public class MVTable extends TableBase {
         changesUntilAnalyze = nextAnalyze <= 0 ? null : new AtomicInteger(nextAnalyze);
         this.store = store;
         this.transactionStore = store.getTransactionStore();
-        this.isHidden = data.isHidden;
-        boolean b = false;
-        for (Column col : getColumns()) {
-            if (DataType.isLargeObject(col.getType().getValueType())) {
-                b = true;
-                break;
-            }
-        }
-        containsLargeObject = b;
         traceLock = database.getTrace(Trace.LOCK);
 
         primaryIndex = new MVPrimaryIndex(database, this, getId(),
@@ -317,104 +290,12 @@ public class MVTable extends TableBase {
         return false;
     }
 
-    private static String getDeadlockDetails(ArrayList<Session> sessions, boolean exclusive) {
-        // We add the thread details here to make it easier for customers to
-        // match up these error messages with their own logs.
-        StringBuilder buff = new StringBuilder();
-        for (Session s : sessions) {
-            Table lock = s.getWaitForLock();
-            Thread thread = s.getWaitForLockThread();
-            buff.append("\nSession ").append(s.toString())
-                    .append(" on thread ").append(thread.getName())
-                    .append(" is waiting to lock ").append(lock.toString())
-                    .append(exclusive ? " (exclusive)" : " (shared)")
-                    .append(" while locking ");
-            int i = 0;
-            for (Table t : s.getLocks()) {
-                if (i++ > 0) {
-                    buff.append(", ");
-                }
-                buff.append(t.toString());
-                if (t instanceof MVTable) {
-                    if (t.isLockedExclusivelyBy(s)) {
-                        buff.append(" (exclusive)");
-                    } else {
-                        buff.append(" (shared)");
-                    }
-                }
-            }
-            buff.append('.');
-        }
-        return buff.toString();
-    }
-
-    @Override
-    public ArrayList<Session> checkDeadlock(Session session, Session clash,
-            Set<Session> visited) {
-        // only one deadlock check at any given time
-        synchronized (MVTable.class) {
-            if (clash == null) {
-                // verification is started
-                clash = session;
-                visited = new HashSet<>();
-            } else if (clash == session) {
-                // we found a circle where this session is involved
-                return new ArrayList<>(0);
-            } else if (visited.contains(session)) {
-                // we have already checked this session.
-                // there is a circle, but the sessions in the circle need to
-                // find it out themselves
-                return null;
-            }
-            visited.add(session);
-            ArrayList<Session> error = null;
-            for (Session s : lockSharedSessions.keySet()) {
-                if (s == session) {
-                    // it doesn't matter if we have locked the object already
-                    continue;
-                }
-                Table t = s.getWaitForLock();
-                if (t != null) {
-                    error = t.checkDeadlock(s, clash, visited);
-                    if (error != null) {
-                        error.add(session);
-                        break;
-                    }
-                }
-            }
-            // take a local copy so we don't see inconsistent data, since we are
-            // not locked while checking the lockExclusiveSession value
-            Session copyOfLockExclusiveSession = lockExclusiveSession;
-            if (error == null && copyOfLockExclusiveSession != null) {
-                Table t = copyOfLockExclusiveSession.getWaitForLock();
-                if (t != null) {
-                    error = t.checkDeadlock(copyOfLockExclusiveSession, clash,
-                            visited);
-                    if (error != null) {
-                        error.add(session);
-                    }
-                }
-            }
-            return error;
-        }
-    }
-
     private void traceLock(Session session, boolean exclusive, TraceLockEvent eventEnum, String extraInfo) {
         if (traceLock.isDebugEnabled()) {
             traceLock.debug("{0} {1} {2} {3} {4}", session.getId(),
                     exclusive ? "exclusive write lock" : "shared read lock", eventEnum.getEventText(),
                     getName(), extraInfo);
         }
-    }
-
-    @Override
-    public boolean isLockedExclusively() {
-        return lockExclusiveSession != null;
-    }
-
-    @Override
-    public boolean isLockedExclusivelyBy(Session session) {
-        return lockExclusiveSession == session;
     }
 
     @Override
@@ -445,26 +326,6 @@ public class MVTable extends TableBase {
                 }
             }
         }
-    }
-
-    @Override
-    public boolean canTruncate() {
-        if (getCheckForeignKeyConstraints() &&
-                database.getReferentialIntegrity()) {
-            ArrayList<Constraint> constraints = getConstraints();
-            if (constraints != null) {
-                for (Constraint c : constraints) {
-                    if (c.getConstraintType() != Constraint.Type.REFERENTIAL) {
-                        continue;
-                    }
-                    ConstraintReferential ref = (ConstraintReferential) c;
-                    if (ref.getRefTable() == this) {
-                        return false;
-                    }
-                }
-            }
-        }
-        return true;
     }
 
     @Override
@@ -642,24 +503,6 @@ public class MVTable extends TableBase {
         }
     }
 
-    private static void addRowsToIndex(Session session, ArrayList<Row> list,
-            Index index) {
-        sortRows(list, index);
-        for (Row row : list) {
-            index.add(session, row);
-        }
-        list.clear();
-    }
-
-    private static void sortRows(ArrayList<? extends SearchRow> list, final Index index) {
-        Collections.sort(list, new Comparator<SearchRow>() {
-            @Override
-            public int compare(SearchRow r1, SearchRow r2) {
-                return index.compareRows(r1, r2);
-            }
-        });
-    }
-
     @Override
     public void removeRow(Session session, Row row) {
         syncLastModificationIdWithDatabase();
@@ -752,16 +595,6 @@ public class MVTable extends TableBase {
     }
 
     @Override
-    public void checkSupportAlter() {
-        // ok
-    }
-
-    @Override
-    public TableType getTableType() {
-        return TableType.TABLE;
-    }
-
-    @Override
     public Index getScanIndex(Session session) {
         return primaryIndex;
     }
@@ -779,25 +612,6 @@ public class MVTable extends TableBase {
     @Override
     public long getMaxDataModificationId() {
         return lastModificationId.get();
-    }
-
-    public boolean getContainsLargeObject() {
-        return containsLargeObject;
-    }
-
-    @Override
-    public boolean isDeterministic() {
-        return true;
-    }
-
-    @Override
-    public boolean canGetRowCount() {
-        return true;
-    }
-
-    @Override
-    public boolean canDrop() {
-        return true;
     }
 
     @Override
@@ -851,11 +665,6 @@ public class MVTable extends TableBase {
         return primaryIndex.getDiskSpaceUsed();
     }
 
-    @Override
-    public void checkRename() {
-        // ok
-    }
-
     /**
      * Get a new transaction.
      *
@@ -864,21 +673,6 @@ public class MVTable extends TableBase {
     Transaction getTransactionBegin() {
         // TODO need to commit/rollback the transaction
         return transactionStore.begin();
-    }
-
-    @Override
-    public Column getRowIdColumn() {
-        if (rowIdColumn == null) {
-            rowIdColumn = new Column(Column.ROWID, Value.LONG);
-            rowIdColumn.setTable(this, SearchRow.ROWID_INDEX);
-            rowIdColumn.setRowId(true);
-        }
-        return rowIdColumn;
-    }
-
-    @Override
-    public String toString() {
-        return getSQL(false);
     }
 
     @Override
